+++ src/tko/game/game_builder.py
@@ -6,6 +6,7 @@ from tko.game.task import Task
 from tko.util.get_md_link import get_md_link
 from tko.util.to_asc import uni_to_asc
 import os
+from icecream import ic
 
 class GameBuilder:
     def __init__(self, filename: str, database_folder: str):